changed ports *again* because L1202 is offline ...

+++ p2p/kill_all.sh
@@ -2,25 +2,28 @@
 
 # Sends a kill signal to:
 
-# L802
-# L803
-# L804
-# L805
-# L806
-# L807 
+# L1201
+# L1203
+# L1204
+# L1205
+# L1206
+# L1207 
 
 # And kills the associated tmux panels 
 
-numberMachines=6
+numberMachines=7
 
 port=51243
-startIP="2"
+startIP="1"
 room="L120"
 
 network="$room$startIP"
 
 for ((i = 1; i < $numberMachines; i++));
 do
+  if [ "$i" -eq "1" ]; then
+      continue
+  fi
   nextIP=$((startIP + i))
   network="$network $room$nextIP"
 done
@@ -29,5 +32,8 @@ java -jar injectEnd.jar $port $network
 
 for ((i = 0; i < $numberMachines; i++));
 do
+  if [ "$i" -eq "1" ]; then
+      continue
+  fi
   tmux kill-pane -t assignment_2:$i
 done
